from branch master
file _sub_vlc_services, [ function StartSlewAnalysis ] , Improved usuability of StartSlewAnalysis, In condition of a song of 262 seconds, and speed variation between 97% of it's speed and 108% of it's speed, randomly on a distribution of 50 randoms generated rates from selected speed variation, it can conserve and repeat the same song and over to acheive a good trust interval. Problems remain in application that rewind the song it does not do it at the beginning and not the end and somewhat between beginning and ending of the application launch. Even having the self trigger based on sample estimated in 'Running Free' timing, it does somewhat close to rewind on time and a bit rewinding to far and thus because application latency is higher on my platform and look like 7 sec to send a command to the vlc-server from OldVlcServer. We do have to consider __old_vlc_rate should send rate variation freely and StartSlewAnalysis too. So initial Buffering the OldVlcServer may delayed twice the command even after it's completion and may risk to not interpret it within it's lifetime. having to schedule to lock or to priorize some command have to be considered. Report during phase sleeping explain somes details and end-report does confirm for this example of setting, song may last between 189 sec and 220 sec.
file _sub_vlc_services, [ function StartSlewAnalysis, section variable declaration ] , StartSlewAnalysis, accept prefixed-parameter SSAPeriodAnalysis, default set to 2. Actual setting was testing with 4 period and perform corretly.

example of vlc-client setting for __old_vlc_rate:
VLCRATEDURATION=20 VLCPRECISETIME=0.05 VLCNICENC=-7 VLCFIXEDCHUNKRATE=True VLCCHUNKRATE=50 VLCRATELOWESTRATE=97 VLCRATEHIGHESTRATE=108 VLCSERVER=127.0.0.1 VLCSRVPORT=5083 VLCSOCKETCLOSESEC=0

example of vlc-server setting using OldVLCServer
OVLCSAppsSockUsingNice=True OVLCSAppsSockNice=-5 OVLCSAppsSockBuffer=128 OVLCSAppsLunch=vlc OVLCSSockAddr=127.0.0.1 OVLCSSockPort=5083