repository : Fnct.D from branch master
file _sub_Structure, [ function CallArgument, section versionCA==0.0.2 ] , enforcing condition toward the __BOOT_STRAP__ tag to be set only if it's not a Sub-function. This case is when using the prefixed-parameter CAIsAddParent is set to True it generate extra information inside __call_locality as 3rd member to be the Parent from where this function was initialized. In case CAIsAddParent=True, it should forbid generating a boot-strap to allow calling ... sub-sub-function which is naturally inaccessible when the top-function was never called. Default prefixed-variable CAIsChildBoot now set to True by default to let appear mistake but will not broke the function and will display which function have to be corrected and what you have to do... Initially the case over 'let the prefixed-variable CAIsAddParent' to master or not the presence of boot-strap is unclear and was holding mistake during parsing. This was corrected and do have to re-correct all function to set CAIsChildBoot to False or the CallArgument==0.0.2 will tell you it was disabled for uselessly create a calling method beyond reachablility.