Update Dpkginsert.pl
Please, the previous version is not correct, start at this version
and you can get access to INSERT. version after will also create a 
insert-PackageInformation.sql as example if user don't own a 
dpkg application it can follow with insert created.
Mostly put back the original function. The OneToOneInsert seems to not evaluate correctly calling to an eval. Insert-file of PackageInformation will also be added as example. 
Take note Fnct.d/src/perl will not merge into install path of the /etc for linux user that try to use it, it can be move and installation script will not add this file inside /etc.