Adding the Summarizing function example with DpkgSelect.pl
Add the sub function and show two query, one based on the post-insertion from DpkgInsert.pl and lead to summarise installed package and de-installed pack The other is giving the number of package not dependant of the processor and the actual processor and how many package. The goal, make a sub function passing the $DBhandler or resulting from DBI->connect(...).