Add DpkgCreate.pl before adding XML configuration.
Hopefully, with DpkgCreate.pl it create the dpkg require database to start the the application to respond to dabase value. Despite addition of Posix-environment variable, adding a XML configuration using whiptail for the graphical interface and example with perl module like XML::Simple, XML::LibXML, XML::Writer. This application can also read SQL file procesing INSERT. But, having DROP table and create table it process in order of DROP -> CREATE -> INSERT. It's actually not able to recognize an insert from a table not create like all the foreign key will not pass until the create table with the foreign key will be processes. Be sure you add your create statement in order of Primary Key / Secondary Key and not create a table with Secondary Key while it refer to unknow table.