The SQL file to create by hand the database for dpkg.
This is just the DROP, CREATE and an update. sqlite_sequence is a protected table should not be erasable from sqlite interface. Erasing it from the Database application in Gnome, I take a note it use extra privilege to allow erasing. In this case you can only do a delete from sqlite_sequence. Using update it update the PkId as primary key for PackageInformation.