PerlCode gnome windows interpreter.
A little dialog windows in gnome allowing a faint and rapid access to process perl from a default file name and doing the interpretation directly if you don't own IDE that support function like hit 'F5' to execute the perl script. This one create a file and do call it.