Adding WhiptailDpkgConf.pl, perl apps configurator.
This is the 2 first windows in text-mode showing promise of Whiptail an efficient console-gui manager fast and rapid when it's the time to make configuration out of Xwindows manager or not depending of Windows/Gnome interface to access to all the data. This application will be responsible to manage a 1 configuration file for all registered application. Actually it sibling application DpkgCreate, DpkgSelect, DpkgSearchDb, DpkgInsert, it allow to run the application with stored variable, skipping in variable to insert from the prompt it will be recorded and evaluated to allow your applications to use latest configuration. Don't commit to this level it only show the screen presentation of https://theysaysomethinginterrestingapparentl.blogspot.com/ .