Update WhiptailDpkgConf.pl
Adding menu for DpkgCreate and its general sub menu. Actually the  back function will require to descend in the menu list to access to it. The DpkgCreate holding couple of Posix environment variable to be held on the prompt in the case it processing an SQL drop/create/insert statement. Change the main $d variable to $TUI (Terminal User Interface), and $SubTUI will also be used when the main TUI is not fitting all the requirement within configuration page.