Adding the tools to validate my c/c++ md5 digest problem.
Between prompt_md5_ssl  and prompt_md5_check which one will get a correct digest return by using the Encryption/Decryption from ssl-layer in java a powerful and easy way to generate key. The algorithm for base64 is not  made yet and a retrieval method in discovered code on a stack-exchange board news with md5 question offer this part. Once compiled, it also solve and tell which code is correct and possible to get the word digested. The C makefile show the example and will be added.