Adding rest of the Arduino SD test.
adding sketch.yaml for arduino-cli user. It include all the specification took out of Arduino-IDE including option to run the loop on Core1 and all Event on Core1. Actually you can compile another arduino project under loop on Core0 and Event on Core0 and will have 2 application running. Usually the same reality when it's the time to pass you by the wired esp32 to a wifi update by OTA which the application receive the update and transfert it to either app0 or app1 space. This is why OTA take place sometime, you can't run an application on the memory region App0 and asking to OTA to write on the App0 in the same, it write on App1 and run it from. 
MD5 header provided by the application it's should be the same as the one inside the library I take it from github too.